+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const listContacts = require("./listContacts");
+const getContactById = require("./getContactById");
+const removeContact = require("./removeContact");
+const addContact = require("./addContact");
+const updateContact = require("./updateContact");
+
+module.exports = {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContact,
+};
