--- models/contacts.js
@@ -1,55 +1,0 @@
-const { randomUUID } = require("crypto");
-const DB = require("../db/db");
-const db = new DB("../models/contacts.json");
-
-const listContacts = async () => {
-  console.log(db);
-  return await db.read();
-};
-
-const getContactById = async (contactId) => {
-  const contacts = await db.read();
-  const [contact] = contacts.filter((contact) => contact.id === contactId);
-  return contact;
-};
-
-const removeContact = async (contactId) => {
-  const contacts = await db.read();
-  const index = contacts.findIndex((contact) => contact.id === contactId);
-  if (index !== -1) {
-    const [contact] = contacts.splice(index, 1);
-    await db.write(contacts);
-    return contact;
-  }
-  return null;
-};
-
-const addContact = async (body) => {
-  const contact = await db.read();
-  const newContact = {
-    id: randomUUID(),
-    ...body,
-  };
-  contact.push(newContact);
-  await db.write(contact);
-  return newContact;
-};
-
-const updateContact = async (contactId, body) => {
-  const contacts = await db.read();
-  const index = contacts.findIndex((contact) => contact.id === contactId);
-  if (index !== -1) {
-    contacts[index] = { ...contacts[index], ...body };
-    await db.write(contacts);
-    return contacts[index];
-  }
-  return null;
-};
-
-module.exports = {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-};
