+++ models/contacts.json
@@ -1,21 +1,9 @@
 [
   {
-    "id": "1",
-    "name": "Allen Raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792"
-  },
-  {
     "id": "2",
     "name": "Chaim Lewis",
     "email": "dui.in@egetlacus.ca",
-    "phone": "(294) 840-6685"
-  },
-  {
-    "id": "5",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(777) 472-5218"
+    "phone": "0777777777"
   },
   {
     "id": "6",
@@ -52,5 +40,11 @@
     "name": "Coret",
     "email": "nibh@semsempererat.com",
     "phone": "0954725218"
+  },
+  {
+    "id": "e7032922-ca4d-4e88-bbca-5d61fdc72b46",
+    "name": "Chaim",
+    "email": "nibh@semsempererat.com",
+    "phone": "0999999999"
   }
 ]
